YAAUM-25 Do some refactoring
