+++ app/src/main/kotlin/dev/yaaum/app/YauumApplication.kt
@@ -1,9 +1,8 @@
 package dev.yaaum.app
 
 import android.app.Application
-import arrow.core.raise.fold
-import dev.yaaum.data.system.timeusage.impl.di.timeUsageSystemModule
-import dev.yaaum.data.system.timeusage.source.TimeUsageDataSource
+import dev.yaaum.data.source.system.timeusage.impl.di.timeUsageSystemModule
+import dev.yaaum.data.source.system.timeusage.source.TimeUsageDataSource
 import dev.yaaum.domain.timeusage.GetStatisticsAboutAllAppsUseCase
 import dev.yaaum.domain.timeusage.impl.di.timeUsageDomainModule
 import dev.yaaum.host.navigation.navigationInit
