+++ data/repository/timeusage/impl/src/main/kotlin/dev/yaaum/repository/timeusage/impl/TimeUsageRepositoryImpl.kt
@@ -6,7 +6,7 @@ import dev.yaaum.data.repository.core.exception.base.BaseRepoException
 import dev.yaaum.data.repository.timeusage.TimeUsageRepository
 import dev.yaaum.data.repository.timeusage.model.TimeUsageRepoModel
 import dev.yaaum.data.repository.timeusage.model.toRepoModel
-import dev.yaaum.data.system.timeusage.source.TimeUsageDataSource
+import dev.yaaum.data.source.system.timeusage.source.TimeUsageDataSource
 
 class TimeUsageRepositoryImpl(
     private val timeUsageDataSource: TimeUsageDataSource,
