+++ data/source/system/core/build.gradle.kts
@@ -1,0 +1,9 @@
+plugins {
+    id("yaaum.convention.feature")
+    id("yaaum.convention.common.detekt")
+    id("yaaum.convention.common.ktlint")
+}
+
+android {
+    namespace = "dev.yaaum.data.source.system.core"
+}
