+++ data/source/system/core/src/main/kotlin/dev/yaaum/data/source/system/core/exception/base/BaseSystemException.kt
@@ -1,0 +1,6 @@
+package dev.yaaum.data.source.system.core.exception.base
+
+/**
+ * Base class to cover all types of exceptions for :data:system level
+ */
+abstract class BaseSystemException : Exception()
