+++ data/source/system/core/src/main/kotlin/dev/yaaum/data/source/system/core/model/base/BaseSystemModel.kt
@@ -1,0 +1,6 @@
+package dev.yaaum.data.source.system.core.model.base
+
+/**
+ * Base class for :data:system level models hierarchy
+ */
+interface BaseSystemModel
