--- data/system/core/build.gradle.kts
@@ -1,9 +1,0 @@
-plugins {
-    id("yaaum.convention.feature")
-    id("yaaum.convention.common.detekt")
-    id("yaaum.convention.common.ktlint")
-}
-
-android {
-    namespace = "dev.yaaum.data.system.core"
-}
