--- data/system/core/src/main/kotlin/dev/yaaum/data/system/core/exception/base/BaseSystemException.kt
@@ -1,6 +1,0 @@
-package dev.yaaum.data.system.core.exception.base
-
-/**
- * Base class to cover all types of exceptions for :data:system level
- */
-abstract class BaseSystemException : Exception()
