--- data/system/core/src/main/kotlin/dev/yaaum/data/system/core/model/base/BaseSystemModel.kt
@@ -1,6 +1,0 @@
-package dev.yaaum.data.system.core.model.base
-
-/**
- * Base class for :data:system level models hierarchy
- */
-interface BaseSystemModel
